+++ examples/light-switch-app/silabs/include/ShellCommands.h
@@ -19,12 +19,8 @@
 
 #pragma once
 
-#if defined(ENABLE_CHIP_SHELL)
-
 namespace LightSwitchCommands {
 
 void RegisterSwitchCommands();
 
 } // namespace LightSwitchCommands
-
-#endif // defined(ENABLE_CHIP_SHELL)
